Fix Bug & Quick Sort choose pivot
1.Insertion sort bug
2.Quick Sort pivot choose experience

random choose, middle choose are both not ok

+++ SortAlgorithms/BasicSort.cpp
@@ -41,7 +41,7 @@ void InsertSortPart(std::vector<int>& sort_vec, unsigned long idx_begin, unsigne
                 break;
             }
         }
-        for (unsigned long start_move = i; start_move > j; start_move--) {
+        for (long long start_move = i; start_move > j; start_move--) {
             sort_vec[start_move] = sort_vec[start_move - 1];
         }
         sort_vec[j + 1] = sort_vec[i];
